Update testing-contracts.md
The second test should revert because addr1 has a balance of 0 and is attempting to transfer 50. But as originally written, the second test does not fail. Line 196 should be "expect(await hardhatToken.connect(addr1).transfer(owner.address, 1)" and not "await expect( hardhatToken.connect(addr1).transfer(owner.address, 1)". As a result, the output also much be revised to reflect the failure.